Make sashimi-esque plot with ggplot2

Shows only junction reads and can optionally show splicing variation across groups.

make_differential_splicing_plot(y, x = numeric(nrow(y)) + 1,
  exons_table = NULL, len = 500, length_transform = function(g) log(g +
  1), main_title = NA, snp_pos = NA, summary_func = colMeans,
  legend_title = "Mean count")

Arguments

y
[samples] x [introns] matrix of counts

x
[samples] vector of group membership. Can be numeric, factor or character.

exons_table
An optional data frame specifying exons, with columns: chr start end strand gene_name. For hg19 see data/gencode19_exons.txt.gz

len
Number of segments each curve is constructed from, controls smoothness.

length_transform
Function controlling how true genomic space is mapped to the plot for improved visability. Use the identity function (i.e. function(g) g

main_title
Title

snp_pos
An optional list of SNP positions to mark

summary_func
Function to summarize counts within groups: usually colMeans or colSums

legend_title
Match this to summary_func